feat: the `Game` struct now iterates moves
