+++ chessie/src/main.rs
@@ -7,8 +7,8 @@ fn main() {
     // let fen = "8/8/4n3/8/3K4/8/8/k7 w - - 0 1";
     // let fen = "3r4/8/8/3N4/3K4/8/8/k7 w - - 0 1";
     // let fen = "4k3/8/8/8/8/8/8/R3K2R w KQ - 0 1"; // Both castling sides are legal
-    let fen = "2n1k3/1P6/8/5pP1/5n2/2P1P3/P7/4K3 w - f6 0 1";
-    // let fen = FEN_STARTPOS;
+    // let fen = "2n1k3/1P6/8/5pP1/5n2/2P1P3/P7/4K3 w - f6 0 1";
+    let fen = FEN_STARTPOS;
     // let fen = "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1";
     // let fen = "r3k2r/p1ppqpb1/b3pnp1/3PN3/1p2P3/2N1nQ1p/PPPB1PPP/R2B1K1R w kq - 4 3"; // Double check
     // let fen = "r3k2r/p1ppqpb1/b3pnp1/3PN3/1p2P3/2N2Q1p/PPPB1PPP/R2B1K1R w kq - 4 3"; // Single check
@@ -21,13 +21,13 @@ fn main() {
     // let mask = pos.pawns(Color::White);
     // let mask = Square::B7.bitboard();
     // let mask = Square::E1.bitboard();
-    let movegen = MoveGen::new(&pos);
+    let game = Game::new(pos);
 
-    eprintln!("{movegen:?}");
+    eprintln!("{game:?}");
 
     eprintln!("Moves:");
     let mut total = 0;
-    for mv in movegen {
+    for mv in game.iter() {
         eprintln!("{mv:?}");
         total += 1;
     }
